FE-514: Move expression IR and SymPy compiler to src/expression
Reorganize into src/expression/{ts-to-ir,ir-to-sympy} and remove the
compile-to-sympy convenience wrapper. The LSP checker now uses
compileToIR directly and reports "Invalid expression" instead of
"SymPy" in diagnostics.
