+++ libs/@hashintel/petrinaut/src/expression/expression-ir.ts
