+++ libs/@hashintel/petrinaut/src/lsp/lib/checker.ts
@@ -4,9 +4,9 @@ import type { SDCPN } from "../../core/types/sdcpn";
 import {
   buildContextForDifferentialEquation,
   buildContextForTransition,
-  compileToSymPy,
-  type SymPyResult,
-} from "../../simulation/simulator/compile-to-sympy";
+  compileToIR,
+  type IRResult,
+} from "../../expression/ts-to-ir/compile-to-ir";
 import type { SDCPNLanguageServer } from "./create-sdcpn-language-service";
 import { getItemFilePath } from "./file-paths";
 
@@ -31,22 +31,22 @@ export type SDCPNCheckResult = {
   isValid: boolean;
   /** TypeScript error diagnostics grouped by item */
   itemDiagnostics: SDCPNDiagnostic[];
-  /** SymPy compilation warning diagnostics (informational, do not affect validity) */
-  sympyDiagnostics: SDCPNDiagnostic[];
+  /** Expression IR diagnostics (unsupported math expressions, do not affect validity) */
+  expressionDiagnostics: SDCPNDiagnostic[];
 };
 
 /**
- * Creates a synthetic ts.Diagnostic from a SymPy compilation error result.
- * Uses category 0 (Warning) since SymPy compilation failures are informational
- * — the TypeScript code may still be valid, just not convertible to SymPy.
+ * Creates a synthetic ts.Diagnostic from an expression IR compilation error.
+ * Uses category 0 (Warning) since the TypeScript code may still be valid,
+ * just not representable as a pure mathematical expression.
  */
-function makeSymPyDiagnostic(
-  result: SymPyResult & { ok: false },
+function makeExpressionDiagnostic(
+  result: IRResult & { ok: false },
 ): ts.Diagnostic {
   return {
     category: 0, // Warning
-    code: 99000, // Custom code for SymPy diagnostics
-    messageText: `SymPy: ${result.error}`,
+    code: 99000,
+    messageText: `Invalid expression: ${result.error}`,
     file: undefined,
     start: result.start,
     length: result.length,
@@ -54,17 +54,17 @@ function makeSymPyDiagnostic(
 }
 
 /**
- * Appends a SymPy diagnostic to the item diagnostics list, merging with
+ * Appends an expression diagnostic to the item diagnostics list, merging with
  * any existing entry for the same item.
  */
-function appendSymPyDiagnostic(
+function appendExpressionDiagnostic(
   itemDiagnostics: SDCPNDiagnostic[],
   itemId: string,
   itemType: ItemType,
   filePath: string,
-  result: SymPyResult & { ok: false },
+  result: IRResult & { ok: false },
 ): void {
-  const diag = makeSymPyDiagnostic(result);
+  const diag = makeExpressionDiagnostic(result);
   const existing = itemDiagnostics.find(
     (di) => di.itemId === itemId && di.itemType === itemType,
   );
@@ -76,20 +76,20 @@ function appendSymPyDiagnostic(
 }
 
 /**
- * Runs SymPy compilation on all SDCPN code expressions and appends
- * any errors as warning diagnostics.
+ * Validates all SDCPN code expressions as mathematical expressions by
+ * compiling them to the expression IR, appending any errors as warnings.
  */
-function checkSymPyCompilation(sdcpn: SDCPN): SDCPNDiagnostic[] {
+function checkExpressions(sdcpn: SDCPN): SDCPNDiagnostic[] {
   const itemDiagnostics: SDCPNDiagnostic[] = [];
   // Check differential equations
   for (const de of sdcpn.differentialEquations) {
     const ctx = buildContextForDifferentialEquation(sdcpn, de.colorId);
-    const result = compileToSymPy(de.code, ctx);
+    const result = compileToIR(de.code, ctx);
     if (!result.ok) {
       const filePath = getItemFilePath("differential-equation-code", {
         id: de.id,
       });
-      appendSymPyDiagnostic(
+      appendExpressionDiagnostic(
         itemDiagnostics,
         de.id,
         "differential-equation",
@@ -102,12 +102,12 @@ function checkSymPyCompilation(sdcpn: SDCPN): SDCPNDiagnostic[] {
   // Check transition lambdas and kernels
   for (const transition of sdcpn.transitions) {
     const lambdaCtx = buildContextForTransition(sdcpn, transition, "Lambda");
-    const lambdaResult = compileToSymPy(transition.lambdaCode, lambdaCtx);
+    const lambdaResult = compileToIR(transition.lambdaCode, lambdaCtx);
     if (!lambdaResult.ok) {
       const filePath = getItemFilePath("transition-lambda-code", {
         transitionId: transition.id,
       });
-      appendSymPyDiagnostic(
+      appendExpressionDiagnostic(
         itemDiagnostics,
         transition.id,
         "transition-lambda",
@@ -129,7 +129,7 @@ function checkSymPyCompilation(sdcpn: SDCPN): SDCPNDiagnostic[] {
         transition,
         "TransitionKernel",
       );
-      const kernelResult = compileToSymPy(
+      const kernelResult = compileToIR(
         transition.transitionKernelCode,
         kernelCtx,
       );
@@ -137,7 +137,7 @@ function checkSymPyCompilation(sdcpn: SDCPN): SDCPNDiagnostic[] {
         const filePath = getItemFilePath("transition-kernel-code", {
           transitionId: transition.id,
         });
-        appendSymPyDiagnostic(
+        appendExpressionDiagnostic(
           itemDiagnostics,
           transition.id,
           "transition-kernel",
@@ -235,12 +235,12 @@ export function checkSDCPN(
     }
   }
 
-  // Run SymPy compilation checks on all code expressions
-  const sympyDiagnostics = checkSymPyCompilation(sdcpn);
+  // Validate expressions as mathematical IR
+  const expressionDiagnostics = checkExpressions(sdcpn);
 
   return {
     isValid: itemDiagnostics.length === 0,
     itemDiagnostics,
-    sympyDiagnostics,
+    expressionDiagnostics,
   };
 }
